+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-lib/src/main/java/com/microsoft/azure/toolkit/intellij/common/component/TreeUtils.java
@@ -19,6 +19,7 @@ import com.intellij.ui.tree.TreeVisitor;
 import com.intellij.util.ui.tree.TreeModelAdapter;
 import com.intellij.util.ui.tree.TreeUtil;
 import com.microsoft.azure.toolkit.ide.common.component.Node;
+import com.microsoft.azure.toolkit.ide.common.favorite.Favorite;
 import com.microsoft.azure.toolkit.ide.common.icon.AzureIcons;
 import com.microsoft.azure.toolkit.intellij.common.IntelliJAzureIcons;
 import com.microsoft.azure.toolkit.intellij.common.action.IntellijAzureActionManager;
@@ -261,12 +262,18 @@ public class TreeUtils {
     }
 
     public static void selectResourceNode(@Nonnull JTree tree, @Nonnull AzComponent resource) {
-        selectResourceNode(tree, resource, false);
+        final DefaultMutableTreeNode r = (DefaultMutableTreeNode) tree.getModel().getRoot();
+        DefaultMutableTreeNode root = ((DefaultMutableTreeNode) r.getChildAt(1)); // apps root
+        if (resource instanceof Favorite) {
+            resource = ((Favorite) resource).getResource();
+            root = (DefaultMutableTreeNode) r.getChildAt(0); // favorite root
+        }
+        if (Objects.nonNull(resource)) {
+            selectResourceNode(tree, resource, root);
+        }
     }
 
-    public static void selectResourceNode(@Nonnull JTree tree, @Nonnull AzComponent resource, boolean underTypeGroup) {
-        final DefaultMutableTreeNode root = (DefaultMutableTreeNode) tree.getModel().getRoot();
-        final DefaultMutableTreeNode node = TreeUtil.findNode(root, underTypeGroup ? TreeUtils::underTypeGroups : TreeUtils::underAppGroups);
+    public static void selectResourceNode(@Nonnull JTree tree, @Nonnull AzComponent resource, DefaultMutableTreeNode node) {
         tree.putClientProperty(HIGHLIGHTED_RESOURCE_KEY, resource);
         Optional.ofNullable(node).ifPresent(n -> TreeUtils.selectNode(tree, new NodeFinder() {
             @Override
@@ -278,7 +285,7 @@ public class TreeUtils {
             public boolean contains(final TreePath path) {
                 final Object current = ((DefaultMutableTreeNode) path.getLastPathComponent()).getUserObject();
                 final ResourceId resourceId = ResourceId.fromString(resource.getId() + "DUMMY");
-                if (underTypeGroup && current instanceof AzService s && s.getName().equalsIgnoreCase(resourceId.providerNamespace())) {
+                if (current instanceof AzService s && s.getName().equalsIgnoreCase(resourceId.providerNamespace())) {
                     return true;
                 }
                 // why append? consider resource `xxx/abc` and `xxx/abcd`
@@ -309,7 +316,7 @@ public class TreeUtils {
         listener.set(new TreeModelAdapter() {
             @Override
             protected void process(@NotNull final TreeModelEvent event, @NotNull final EventType type) {
-                if (event.getTreePath().equals(checkpoint.get()) && type != EventType.NodesRemoved) {
+                if (event.getTreePath().equals(checkpoint.get()) && type != EventType.NodesRemoved && type != EventType.NodesChanged) {
                     doSelectNode(tree, finder, checkpoint, listener.get());
                 }
             }
@@ -337,7 +344,10 @@ public class TreeUtils {
             }
         }).onSuccess(path -> {
             tree.getModel().removeTreeModelListener(listener);
-            TreeUtil.selectPath(tree, path, true);
+        }).onProcessed(path -> {
+            if (Objects.nonNull(checkpoint.get())) {
+                TreeUtil.selectPath(tree, checkpoint.get(), true);
+            }
         });
     }
 }