+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/code/spring/YamlValueCompletionProvider.java
@@ -5,6 +5,7 @@
 
 package com.microsoft.azure.toolkit.intellij.connector.code.spring;
 
+import com.google.common.collect.ImmutableMap;
 import com.intellij.codeInsight.completion.CompletionParameters;
 import com.intellij.codeInsight.completion.CompletionProvider;
 import com.intellij.codeInsight.completion.CompletionResultSet;
@@ -67,14 +68,16 @@ public class YamlValueCompletionProvider extends CompletionProvider<CompletionPa
             return;
         }
         final List<? extends Resource<?>> resources = definitions.stream()
-                .flatMap(d -> Utils.listResourceForDefinition(module.getProject(), d).stream())
+                .flatMap(d -> Utils.checkCancelled(() -> d.getResources(module.getProject())).stream())
                 .toList();
         ProgressManager.checkCanceled();
-        resources.stream()
+        final List<LookupElement> elements = resources.stream()
                 .map(resource -> createYamlValueLookupElement(module, resource))
                 .filter(Objects::nonNull)
-                .forEach(result::addElement);
+                .collect(Collectors.toList());
+        elements.forEach(result::addElement);
         AzureTelemeter.log(AzureTelemetry.Type.OP_END, OperationBundle.description("boundary/connector.complete_values_in_yaml"));
+        AzureTelemeter.info("connector.resources_count.yaml_value_code_completion", ImmutableMap.of("count", elements.size() + "", "key", key));
     }
 
     @Nullable
@@ -85,8 +88,8 @@ public class YamlValueCompletionProvider extends CompletionProvider<CompletionPa
                 .withIcon(IntelliJAzureIcons.getIcon(StringUtils.firstNonBlank(definition.getIcon(), AzureIcons.Common.AZURE.getIconPath())))
                 .withBoldness(true)
                 .withPresentableText(resource.getName())
-                .withTypeText("String")
-                .withTailText(" " + ((AzResource) resource.getData()).getResourceTypeName())
+                .withTypeText(((AzResource) resource.getData()).getResourceTypeName())
+                .withTailText(" " + ((AzResource) resource.getData()).getResourceGroupName())
                 .withLookupStrings(Arrays.asList(resource.getName(), ((AzResource) resource.getData()).getResourceGroupName()))
                 .withInsertHandler((context, item) -> handleYamlConnection(azureModule, resource, context));
     }
@@ -111,7 +114,7 @@ public class YamlValueCompletionProvider extends CompletionProvider<CompletionPa
         final PsiElement element = PsiUtil.getElementAtOffset(context.getFile(), context.getStartOffset());
         final YAMLPsiElement yamlElement = Objects.requireNonNull(PsiTreeUtil.getParentOfType(element, YAMLPsiElement.class));
         final String key = YAMLUtil.getConfigFullName(yamlElement);
-        final SpringSupported<?> definition = (SpringSupported<?>)connection.getResource().getDefinition();
+        final SpringSupported<?> definition = (SpringSupported<?>) connection.getResource().getDefinition();
         final String value = definition.getSpringProperties(key).stream()
                 .filter(pair -> StringUtils.equalsIgnoreCase(pair.getKey(), key))
                 .findFirst()