+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-resource-connector-lib-java/src/main/java/com/microsoft/azure/toolkit/intellij/connector/dotazure/BeforeRunTaskAdder.java
@@ -32,12 +32,15 @@ public class BeforeRunTaskAdder implements RunManagerListener, ConnectionTopics.
             return;
         }
         final List<Connection<?, ?>> connections = profile.getConnections();
-        final List<BeforeRunTask<?>> tasks = new ArrayList<>(config.getBeforeRunTasks());
-        tasks.removeIf(t -> t instanceof DotEnvBeforeRunTaskProvider.LoadDotEnvBeforeRunTask);
-        if (connections.stream().anyMatch(c -> c.isApplicableFor(config))) {
-            tasks.add(new DotEnvBeforeRunTaskProvider.LoadDotEnvBeforeRunTask(config));
+        synchronized (config.getBeforeRunTasks()) {
+            final List<BeforeRunTask<?>> tasks = new ArrayList<>(config.getBeforeRunTasks().stream()
+                .filter(task -> !(task instanceof DotEnvBeforeRunTaskProvider.LoadDotEnvBeforeRunTask))
+                .toList());
+            if (connections.stream().anyMatch(c -> c.isApplicableFor(config))) {
+                tasks.add(new DotEnvBeforeRunTaskProvider.LoadDotEnvBeforeRunTask(config));
+            }
+            config.setBeforeRunTasks(tasks);
         }
-        config.setBeforeRunTasks(tasks);
     }
 
     @Override